dded client-side interceptor to automatically add metadata to service call
Updated README to reflect the change
